+++ CHANGES.md
@@ -20,6 +20,7 @@
   * A ring-route middleware `calfpath.route/ring-handler-middleware` to wrap Ring handlers into route handlers
 * Helper fn to build routes from given route specs
 * Allow non-literal string URI-patterns in `calfpath.core/->uri`
+* Fix `calfpath.route/update-fallback-400` to add fallback 400 route on one or more URI entry, instead of all
 * [TODO] BREAKING CHANGE: Drop `calfpath.core/make-uri-handler` in favor of Calfpath routes